[core] Add Url class
Passing url-string / page-string / query-string / fragment-string around separately is better than passing random untyped strings around... but when a whole tree of functions is based on passing a page-string around, and then I want to pass a page-string _and_ a fragment-string... argh.

Having an object which collects page/query/fragment together and passes them around as a unit is a big rework, but makes a lot of things Just Work that were previously impractical.